requirements.txt

# Use modular LangChain packages (install latest available)
langchain-core
langchain-openai
python-dotenv
openai
tiktoken
